+++ services/gateway/sources/version2/fees.js
@@ -15,7 +15,7 @@
  */
 module.exports = {
 	type: 'moleculer',
-	method: 'fees.fee.estimates',
+	method: 'fees.estimates',
 	params: {},
 	definition: {
 		data: {
